+++ src/tool/hpcrun/sample-sources/blame-shift/blame-shift.h
@@ -3,7 +3,7 @@
 
 #include <cct/cct.h>
 
-typedef void (*bs_fn_t)(void* arg, int metric_id, cct_node_t* node, int metric_incr);
+typedef void (*bs_fn_t)(void* arg, int metric_id, cct_node_t* node, float metric_incr);
 
 typedef struct bs_fn_entry_t {
   struct bs_fn_entry_t* next;
@@ -18,7 +18,7 @@ typedef enum bs_type{
 
 
 void blame_shift_register(bs_fn_entry_t* entry);
-void blame_shift_apply(int metric_id, cct_node_t* node, int metric_incr);
+void blame_shift_apply(int metric_id, cct_node_t* node, float metric_incr);
 void blame_shift_source_register(bs_type bst);
 int blame_shift_source_available(bs_type bst);
