+++ src/tool/hpcrun/sample-sources/blame-shift/undirected.h
@@ -39,7 +39,7 @@ void undirected_blame_idle_begin(undirected_blame_info_t *bi);
 void undirected_blame_idle_end(undirected_blame_info_t *bi);
 
 void undirected_blame_sample(void *arg, int metric_id, 
-			     cct_node_t *node, int metric_incr);
+			     cct_node_t *node, float metric_incr);
 
 
 #endif // __undirected_h__
